first from_awkward test
